refactor: moving the default header into a default-header folder
Also cleaning up some other files in the header folder, like ‘generic/messages.js’ which I folded into the learning header’s messages.
