+++ shell/header/default-header/Menu/Menu.jsx
@@ -1,6 +1,6 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import { CSSTransition } from 'react-transition-group';
-import PropTypes from 'prop-types';
 
 const MenuTrigger = ({ tag, className, ...attributes }) => React.createElement(tag, {
   className: `menu-trigger ${className}`,
@@ -273,4 +273,4 @@ Menu.defaultProps = {
   transitionClassName: 'menu-content',
 };
 
-export { Menu, MenuTrigger, MenuContent };
+export { Menu, MenuContent, MenuTrigger };
