--- shell/header/test-utils.js
@@ -1,6 +1,0 @@
-const executeThunk = async (thunk, dispatch, getState) => {
-  await thunk(dispatch, getState);
-  await new Promise(setImmediate);
-};
-
-export default executeThunk;
